Drop pytest plugin from Red's package
