+++ Makefile
@@ -39,7 +39,7 @@ stylediff:
 
 # Translations
 gettext:
-	$(PYTHON) -m redgettext --command-docstrings --verbose --recursive redbot --exclude-files "redbot/pytest/**/*"
+	$(PYTHON) -m redgettext --command-docstrings --verbose --recursive redbot
 upload_translations:
 	crowdin upload sources
 download_translations:
