+++ tests/cogs/test_alias.py
@@ -1,5 +1,14 @@
+from redbot.cogs.alias import Alias
+from redbot.core import Config
+
 import pytest
-from redbot.pytest.alias import *
+
+
+@pytest.fixture()
+def alias(config, monkeypatch):
+    with monkeypatch.context() as m:
+        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
+        return Alias(None)
 
 
 def test_is_valid_alias_name(alias):
