+++ tests/cogs/test_mod.py
@@ -1,7 +1,19 @@
 from collections import namedtuple
+
+from redbot.core import modlog
+
 import pytest
 
-from redbot.pytest.mod import *
+
+@pytest.fixture
+async def mod(config, monkeypatch, red):
+    from redbot.core import Config
+
+    with monkeypatch.context() as m:
+        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
+
+        await modlog._init(red)
+        return modlog
 
 
 async def test_modlog_register_casetype(mod):
