+++ tests/core/_downloader/.gitattributes
