+++ tests/core/_downloader/fixtures.py
