+++ tests/core/test_data_manager.py
@@ -3,8 +3,25 @@ from pathlib import Path
 
 import pytest
 
-from redbot.pytest.data_manager import *
 from redbot.core import data_manager
+
+
+@pytest.fixture(autouse=True)
+def cleanup_datamanager():
+    data_manager.basic_config = None
+
+
+@pytest.fixture()
+def data_mgr_config(tmpdir):
+    default = data_manager.basic_config_default.copy()
+    default["BASE_DIR"] = str(tmpdir)
+    return default
+
+
+@pytest.fixture()
+def cog_instance():
+    thing = type("CogTest", (object,), {})
+    return thing()
 
 
 def test_no_basic(cog_instance):
