+++ tools/edit_testrepo.py
@@ -12,7 +12,7 @@ edit_testrepo.py import:
     imports the repo from ``redbot/tests/downloader_testrepo.export`` using
     git's fast-import command and updates repo's working tree.
 edit_testrepo.py export:
-    It exports repo from provided directory into ``redbot/tests/downloader_testrepo.export``
+    It exports repo from provided directory into ``tests/core/_downloader/downloader_testrepo.export``
     using git's fast-export. To make the file more useful for developers,
     it's called with option that adds extra directive ``original-oid <SHA1SUM>``,
     which while ignored by import, might ease up creating tests without importing the repo.
@@ -42,7 +42,9 @@ import click
 
 
 MAIN_DIRECTORY = Path(__file__).absolute().parent.parent
-TEST_REPO_EXPORT_PTH: Path = MAIN_DIRECTORY / "redbot" / "pytest" / "downloader_testrepo.export"
+TEST_REPO_EXPORT_PTH: Path = (
+    MAIN_DIRECTORY / "tests" / "core" / "_downloader" / "downloader_testrepo.export"
+)
 
 
 class ClickCustomPath(click.Path):
